# Add concurrency to the fetchData function