--- a/uchiwa/daemon/datacenters.go
+++ b/uchiwa/daemon/datacenters.go
@@ -2,7 +2,7 @@ package daemon
 
 import "github.com/sensu/uchiwa/uchiwa/structs"
 
-func (d *Daemon) buildDatacenter(name *string, info *structs.Info) *structs.Datacenter {
+func (f *DatacenterFetcher) buildDatacenter(name *string, info *structs.Info) *structs.Datacenter {
 	datacenter := structs.Datacenter{
 		Name:    *name,
 		Info:    *info,